requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml>=5.3.0
markdownify==0.13.1
langchain-openai>=0.3.14
langchain-core>=0.3.55
langgraph>=0.3.18
langchain-tavily>=0.1.6
gspread==6.1.4
google-auth==2.37.0
apscheduler==3.10.4
python-dotenv==1.0.1